move feedback reducer to reducers folder
